% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.r
\name{gaussian_singleValue}
\alias{gaussian_singleValue}
\title{density of a multivariate normal distribution for on column vector}
\usage{
gaussian_singleValue(
  x,
  mu,
  varcovM,
  compute_mat = TRUE,
  determ = NULL,
  inv = NULL
)
}
\arguments{
\item{x}{matrix : one column}

\item{mu}{mean (vector of mean)}

\item{varcovM}{matrix - square matrix (nrow == ncol) strictly positive}

\item{compute_mat}{Boolean - True to compute inverse and det}

\item{determ}{if !compute_mat determinant of varcovM (output of det(varcovM))}

\item{inv}{if !compute_mat inverse of varcovM (output of solve(varcovM))}
}
\value{
vector
}
\description{
Description
}
\details{
Details (opt)
}
