% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heavyRpackage-package.R
\docType{package}
\name{heavyRpackage-package}
\alias{heavyRpackage}
\alias{heavyRpackage-package}
\title{heavyRpackage: What the Package Does (One Line, Title Case)}
\description{
Documentation à ajouter, tel fonction fait ceci
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
