% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/mvnpdf.r
\name{mvnpdfoptimC}
\alias{mvnpdfoptimC}
\alias{mvnpdfsmartC}
\alias{mvnpdf}
\alias{mvnpdfsmart}
\alias{mvnpdfoptim_par}
\alias{mvnpdfoptim_parIter}
\alias{mvnpdfsmart_par}
\alias{mvnpdfsmart_sleepy}
\alias{mvnpdfsmart_sleepy_par}
\title{density of a multivariate normal distribution on Rp at n points}
\usage{
mvnpdfoptimC(x, mean, varcovM, Log = TRUE)

mvnpdfsmartC(x, mean, varcovM, Log = TRUE)

mvnpdf(x, varcovM = diag(nrow(x)), mu = rep(0, nrow(varcovM)), Log = FALSE)

mvnpdfsmart(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfoptim_par(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfoptim_parIter(
  x,
  mean = rep(0, nrow(x)),
  varcovM = diag(nrow(x)),
  Log = TRUE,
  ncores = 1
)

mvnpdfsmart_par(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfsmart_sleepy(
  x,
  mean = rep(0, nrow(x)),
  varcovM = diag(nrow(x)),
  Log = TRUE
)

mvnpdfsmart_sleepy_par(
  x,
  mean = rep(0, nrow(x)),
  varcovM = diag(nrow(x)),
  Log = TRUE
)
}
\arguments{
\item{x}{matrice, observations: point pour lequel on
veut calculer la densité de loi normale}

\item{varcovM}{matrix - square matrix (nrow == ncol) strictly positive variance covariance matrix}

\item{Log}{Bool - return log of y or not}

\item{mu}{mean (mean vector of a gaussian distribution)}

\item{ncores}{number of parallel cores to split task over when using itertools}
}
\value{
vector of size nrow(x) and x
}
\description{
density of a multivariate normal distribution on Rp at n points
}
\examples{
mat = matrix(c(1,0,0,1), nrow=2, byrow = TRUE)
x=matrix(c(-0.5,1.5,0,1,-1, 1),nrow = 2)
mu = matrix(c(0,0),nrow = 2)
mvnpdf(x=x, varcovM = mat, mu=mu, Log=FALSE)
\dontrun{
n <- 10000
mb <- microbenchmark::microbenchmark(
  mvtnorm::dmvnorm(matrix(1.96, nrow = n, ncol = 2)),
  mvnpdfsmart(x=matrix(1.96, nrow = 2, ncol = n), Log=FALSE),
  mvnpdfoptim_par(x=matrix(1.96, nrow = 2, ncol = n), Log=FALSE),
  times=10L)
mb
}

}
