% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heavyRpackage-package.R
\docType{package}
\name{heavyRpackage-package}
\alias{heavyRpackage}
\alias{heavyRpackage-package}
\title{heavyRpackage: What the Package Does (One Line, Title Case)}
\description{
Documentation à ajouter, tel fonction fait ceci
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/acolajanni/heavyRpackage}
  \item \url{https://acolajanni.github.io/heavyRpackage/}
  \item Report bugs at \url{https://github.com/acolajanni/heavyRpackage/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
