% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.r
\name{mvnpdf}
\alias{mvnpdf}
\title{density of a multivariate normal distribution on Rp at n points}
\usage{
mvnpdf(x, varcovM = diag(nrow(x)), mu = rep(0, nrow(varcovM)), Log = FALSE)
}
\arguments{
\item{x}{matrice, observations: point pour lequel on
veut calculer la densité de loi normale}

\item{varcovM}{matrix - square matrix (nrow == ncol) strictly positive variance covariance matrix}

\item{mu}{mean (mean vector of a gaussian distribution)}

\item{Log}{Bool - return log of y or not}
}
\value{
vector of size nrow(x) and x
}
\description{
density of a multivariate normal distribution on Rp at n points
}
\examples{
mat = matrix(c(1,0,0,1), nrow=2, byrow = TRUE)
x=matrix(c(-0.5,1.5,0,1,-1, 1),nrow = 2)
mu = matrix(c(0,0),nrow = 2)
mvnpdf(x=x, varcovM = mat, mu=mu, Log=FALSE)
}
